Add: `useTokenTotalBalance` utility hook
