+++ src/hooks/index.ts
@@ -50,6 +50,7 @@ export { default as useExtensionsData } from './useExtensionsData';
 export * from './useCanInteractWithColony';
 export { default as useColonyFundsClaims } from './useColonyFundsClaims';
 export { default as useCurrentSelectedToken } from './useCurrentSelectedToken';
+export { default as useTokenTotalBalance } from './useTokenTotalBalance';
 
 /* Used in cases where we need to memoize the transformed output of any data.
  * Transform function has to be pure, obviously
